move @Bean producers into separate classes
